removed handlebar routes folder and moved content into home-routes folder
