+++ controllers/home-routes.js
@@ -3,21 +3,19 @@ const path = require("path");
 const { User, Pairing, Comment, Review } = require("../models");
 const sequelize = require("sequelize");
 
-
 //getting all pairing data for carousel cards
-router.get('/', async (req, res) => {
+router.get("/", async (req, res) => {
   try {
-      const pairingData = await Pairing.findAll({
-        attributes: {exclude: ['user_id']},
-          include: [
-            {
-              model: Review,
-              attributes: {exclude: ['pairing_id', 'user_id']},
-              include: [{model: User, attributes: {exclude: ['id']}}]
-            }
-          ]
-      });
-
+    const pairingData = await Pairing.findAll({
+      attributes: { exclude: ["user_id"] },
+      include: [
+        {
+          model: Review,
+          attributes: { exclude: ["pairing_id", "user_id"] },
+          include: [{ model: User, attributes: { exclude: ["id"] } }],
+        },
+      ],
+    });
 
     const pairings = await pairingData.map((pairing) =>
       pairing.get({ plain: true })
@@ -29,21 +27,44 @@ router.get('/', async (req, res) => {
 });
 
 //once a user is logged in, redirect to home page
-router.get('/login', (req, res) => {
-  if(req.session.loggedIn) {
-      res.redirect('/');
-      return; 
+router.get("/login", (req, res) => {
+  if (req.session.loggedIn) {
+    res.redirect("/");
+    return;
   }
 });
 
 //rendering sign up page
-//once a user clicks sign up or generate pairing will redirect them to the pairings home page 
-//with a modal 
-router.get('/signup', (req, res) => {
+//once a user clicks sign up or generate pairing will redirect them to the pairings home page
+//with a modal
+router.get("/signup", (req, res) => {
   // res.redirect('signup')
 });
 
+// Handlebar routes
 
+router.get("/views/aboutUs", (req, res) => {
+  res.render("aboutUs", { loggedIn: req.session.loggedIn });
+});
 
+router.get("/", (req, res) => {
+  const imagesArr = [{ number: 2 }, { number: 3 }, { number: 4 }];
+  res.render("home", {
+    images: imagesArr,
+    loggedIn: req.session.loggedIn,
+  });
+});
+
+router.get("/views/login", (req, res) => {
+  res.render("login", { loggedIn: req.session.loggedIn });
+});
+
+router.get("/views/contactUs", (req, res) => {
+  res.render("contactUs", { loggedIn: req.session.loggedIn });
+});
+
+router.get("/views/pairing", (req, res) => {
+  res.render("pairing", { loggedIn: req.session.loggedIn });
+});
 
 module.exports = router;
