+++ package.json
@@ -1,20 +1,1 @@
-{
-"scripts": {
-  "start": "node server.js",
-  "dev": "nodemon server.js",
-  "seed": "node seeds/index.js"
-},
-  "dependencies": {
-    "bcrypt": "^5.1.0",
-    "connect-session-sequelize": "^7.1.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "express-handlebars": "^6.0.7",
-    "express-session": "^1.17.3",
-    "mysql2": "^3.1.1",
-    "sequelize": "^6.28.0"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.20"
-  }
-}
+{"scripts":{"start":"node server.js","dev":"nodemon server.js","seed":"node seeds/index.js"},"dependencies":{"@planprice/funnies":"^1.0.0","bcrypt":"^5.1.0","connect-session-sequelize":"^7.1.5","dotenv":"^16.0.3","express":"^4.18.2","express-handlebars":"^6.0.7","express-session":"^1.17.3","mysql2":"^3.1.1","sequelize":"^6.28.0"},"devDependencies":{"nodemon":"^2.0.20"}}
