+++ public/css/style.css
@@ -8,7 +8,6 @@
   -o-transition: all ease-out 0.25s;
   transition: all ease-out 0.25s;
 }
-
 /*  ----Variables---- */
 
 :root {
@@ -29,6 +28,16 @@
   --align-items-center: center;
   --display-flex: flex;
   --cursor-pointer: pointer;
+}
+
+.loader-container {
+  width: 100%;
+  height: 100vh;
+  position: fixed;
+  background: #000
+    url("https://media.giphy.com/media/8agqybiK5LW8qrG3vJ/giphy.gif") center
+    no-repeat;
+  z-index: 1;
 }
 
 a {
