+++ views/layouts/main.handlebars
@@ -17,6 +17,12 @@
     />
     <title>Beer 'n' Bites</title>
   </head>
+{{!-- Loading popup for funnies --}}
+  <div class="loader-container">
+    <div class="funny"></div>
+</div>
+
+
   <body>
     <header>
       <a href="/" class="logoImg"><img
